+++ src/api/v1/Path2D.test.js
@@ -1,12 +1,11 @@
 const defaultPathGeometry = require('./defaultPathGeometry')
-const path = require('../../geometry/path')
 const mat4 = require('../../math/mat4')
 const test = require('ava')
 const Path2D = require('./Path2D')
 const call = require('../registry/call')
 
 // Hook up geometry/path as the implementation for the v1 api to use by default.
-call(defaultPath, path)
+call(defaultPathGeometry, require('../../geometry/path'))
 
 test('Empty path has no points', t => {
   t.deepEqual(new Path2D().getPoints(), [])
